+++ helpers/ApiMarkdownTrait.php
@@ -131,4 +131,50 @@ trait ApiMarkdownTrait
     {
         return $block[1];
     }
+
+    /**
+     * Consume lines for a blockquote element
+     */
+    protected function consumeQuote($lines, $current)
+    {
+        $block = parent::consumeQuote($lines, $current);
+
+        $blockTypes = [
+            'warning',
+            'note',
+            'info',
+            'tip',
+        ];
+
+        // check whether this is a special Info, Note, Warning, Tip block
+        $content = $block[0]['content'];
+        $first = reset($content);
+        if (isset($first[0]) && $first[0] === 'paragraph') {
+            $parfirst = reset($first['content']);
+            if (isset($parfirst[0]) && $parfirst[0] === 'text') {
+                foreach ($blockTypes as $type) {
+                    if (strncasecmp("$type: ", $parfirst[1], $len = strlen($type) + 2) === 0) {
+                        $block[0]['content'][0]['content'][0][1] = $this->translateBlockType($type) . substr($parfirst[1], $len);
+                        $block[0]['blocktype'] = $type;
+                        break;
+                    }
+                }
+            }
+        }
+        return $block;
+    }
+
+    protected abstract function translateBlockType($type);
+
+    /**
+     * Renders a blockquote
+     */
+    protected function renderQuote($block)
+    {
+        $class = '';
+        if (isset($block['blocktype'])) {
+            $class = ' class="' . $block['blocktype'] . '"';
+        }
+        return "<blockquote{$class}>" . $this->renderAbsy($block['content']) . "</blockquote>\n";
+    }
 }