+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Yii Framework 2 apidoc extension Change Log
 - Enh #13: Added a JSON template to output the class structure as a JSON file (tom--)
 - Enh: Added callback `afterMarkdownProcess()` to HTML Guide renderer (cebe)
 - Enh: Added `getHeadings()` method to ApiMarkdown class (cebe)
+- Enh: Added css class to Info, Warning, Note and Tip blocks (cebe)
 - Chg #31: Hightlight.php library is now used for code highlighing, the builtin ApiMarkdown::hightligh() function is not used anymore (cebe)